admin can get jobApp's information from jobPost

+++ src/main/java/com/coficab/app_recrutement_api/jobPost/JobPostRepository.java
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Repository;
 
 import java.util.List;
 
-public interface JobPostRepository extends JpaRepository<JobPost, Integer>, JpaSpecificationExecutor<JobPost> {
+public interface JobPostRepository extends JpaRepository<JobPost, Long>, JpaSpecificationExecutor<JobPost> {
     @Query("""
             SELECT jobPost
             FROM JobPost jobPost
